requirements.txt
Flask~=2.2.2
kafka-python~=2.0.2
pymongo~=4.3.3
requests~=2.28.1
networkx~=3.0
pandas~=1.5.3
matplotlib~=3.6.3